+++ NativeCollections/NativeCollections/NativeConcurrentChunkedStream.cs
@@ -521,7 +521,7 @@ namespace NativeCollections
                             }
 
                             --handle->Chunks;
-                            Unsafe.CopyBlockUnaligned(ref Unsafe.AddByteOffset(ref reference, byteCount), ref *handle->Head->Array, (uint)size);
+                            Unsafe.CopyBlockUnaligned(ref Unsafe.AddByteOffset(ref reference, (nint)byteCount), ref *handle->Head->Array, (uint)size);
                             byteCount += size;
                         }
 
@@ -541,7 +541,7 @@ namespace NativeCollections
                             }
 
                             --handle->Chunks;
-                            Unsafe.CopyBlockUnaligned(ref Unsafe.AddByteOffset(ref reference, byteCount), ref *handle->Head->Array, (uint)remaining);
+                            Unsafe.CopyBlockUnaligned(ref Unsafe.AddByteOffset(ref reference, (nint)byteCount), ref *handle->Head->Array, (uint)remaining);
                         }
                     }
                 }
@@ -588,7 +588,7 @@ namespace NativeCollections
                         }
 
                         --handle->Chunks;
-                        Unsafe.CopyBlockUnaligned(ref Unsafe.AddByteOffset(ref reference, byteCount), ref *handle->Head->Array, (uint)size);
+                        Unsafe.CopyBlockUnaligned(ref Unsafe.AddByteOffset(ref reference, (nint)byteCount), ref *handle->Head->Array, (uint)size);
                         byteCount += size;
                     }
 
@@ -608,7 +608,7 @@ namespace NativeCollections
                         }
 
                         --handle->Chunks;
-                        Unsafe.CopyBlockUnaligned(ref Unsafe.AddByteOffset(ref reference, byteCount), ref *handle->Head->Array, (uint)remaining);
+                        Unsafe.CopyBlockUnaligned(ref Unsafe.AddByteOffset(ref reference, (nint)byteCount), ref *handle->Head->Array, (uint)remaining);
                     }
 
                     handle->ReadOffset = remaining;
@@ -668,7 +668,7 @@ namespace NativeCollections
                     handle->Tail->Next = chunk;
                     handle->Tail = chunk;
                     ++handle->Chunks;
-                    Unsafe.CopyBlockUnaligned(ref *handle->Tail->Array, ref Unsafe.AddByteOffset(ref reference, byteCount), (uint)size);
+                    Unsafe.CopyBlockUnaligned(ref *handle->Tail->Array, ref Unsafe.AddByteOffset(ref reference, (nint)byteCount), (uint)size);
                     byteCount += size;
                 }
 
@@ -688,7 +688,7 @@ namespace NativeCollections
                     handle->Tail->Next = chunk;
                     handle->Tail = chunk;
                     ++handle->Chunks;
-                    Unsafe.CopyBlockUnaligned(ref *handle->Tail->Array, ref Unsafe.AddByteOffset(ref reference, byteCount), (uint)remaining);
+                    Unsafe.CopyBlockUnaligned(ref *handle->Tail->Array, ref Unsafe.AddByteOffset(ref reference, (nint)byteCount), (uint)remaining);
                 }
 
                 handle->WriteOffset = remaining;
